rename context files using kebab case
